Add csvargs command. This is a very rough draft.
csvargs is inspired by xargs where the columns of a csv are used as
arguments for a command. Right now the command is specified as a
string with the --comand argument. The placement of columns in the
command is indicated by "__"s in the command string. The columns are
assumed to appear in the desired order (for example, by using csvcut),
so no provision is made for specifying another order.